Command parity to vanilla (#505)
* Updated show_entity

Changed a little bit the show_entity function from HoverEvents to better usability

* Added to_string to EntityType

Added to_str and to_string functions to allow proper use of identifiers in the show_entity of a TextComponent.

* Clear texts parity

Added translations to clear command, also added to show the player which items are cleaned, and tell something to the user like vanilla.

* Pumpkin command format

Formatted the pumpkin command to show better the text.

* Kill texts parity

Added translations to kill command, also added to show the player who is killed, and tell something to that user like vanilla. Also changed the color to white like vanilla.

* Added translate key of item

Added a function to item_registry which returns the appropriate TextComponent::translate for a item

* Give texts parity

Added translations to give command item, also added to show the player who receive the item, and tell something to that user like vanilla.

* CLIPPY MY LORD

* Reverted entity_type changes

I had not thought that I could use debug to get the name of the type, my mistake

* Added send_message to Player

Added the function send_message to allow sending DisgusedChatMessages directly to a specific player.

* Added msg command

Added msg (tell) command to match with the applied parity changes.

* Bossbar texts parity

Added translations and parity props for Bossbar command texts, without changing their behavior.